move core submit modal to its own component
